requirements.txt
Flask==3.0.0
google-genai>=0.2.2
Werkzeug==3.0.1
